requirements.txt
Django==3.0.7
wheel==0.33.6
pytz==2018.4
Pillow>=6.2.2
gunicorn==19.9.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
django-environ==0.4.5
django-allauth==0.41.0
drf-yasg==1.17.0
django-timezone-field==4.0
django-resized==0.3.11
django-cors-headers==3.3.0
dj-rest-auth==1.1.1
django-extra-settings==0.1.1